Replace around with after plugin
